+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 ### Added
 
+* add references to Inputs and Outputs to enable completion
+  and jump-to-definition
+
 ### Changed
 
 ### Deprecated